+++ RELEASE_NOTES.md
@@ -5,6 +5,18 @@ PresentationKit will use semver after 1.0.
 Until then, breaking changes can happen in any version, and deprecations may be removed in any minor version bump.
 
 
+## 0.2
+
+This version adds new navigation utilities.
+
+### ✨ New Features
+
+* `NavigationButton` can be used to render a `Button` as a `NavigationLink`.
+* `NavigationChevron` can be used to mimic a native navigation chevron.
+* `NavigationContext` can be used to manage an observable navigation path.
+
+
+
 ## 0.1
 
 This is the first public release of PresentationKit.